Added logic to show plots on top of other plots as they are opened. Some refactor
